+++ CHANGES.md
@@ -15,6 +15,7 @@ Features
 * [#1191](https://github.com/java-native-access/jna/pull/1191): Add `c.s.j.p.win32.Advapi32Util#getTokenPrimaryGroup` - [@dbwiddis](https://github.com/dbwiddis).
 * [#1182](https://github.com/java-native-access/jna/pull/1182): Add `toString` to classes extending `c.s.j.ptr.ByReference` - [@dbwiddis](https://github.com/dbwiddis).
 * [#1194](https://github.com/java-native-access/jna/pull/1194): Add `GetConsoleScreenBufferInfo`, `ReadConsoleInput` and `WriteConsole` with associated structures to `c.s.j.p.win32.Wincon` - [@rednoah](https://github.com/rednoah).
+* [#1202](https://github.com/java-native-access/jna/pull/1202): Add `size_t`, `ssize_t`, and `off_t` to `c.s.j.p.unix.LibCAPI` - [@dbwiddis](https://github.com/dbwiddis).
 
 Bug Fixes
 ---------